model_utils.py
import pandas as pd
import numpy as np
from sklearn.metrics import roc_curve, precision_recall_curve, confusion_matrix

def preprocess_patient_data(data):
    """
    Preprocess patient data for model input
    
    Parameters:
    data (pd.DataFrame): Raw patient data
    
    Returns:
    pd.DataFrame: Processed data ready for model input
    """
    # This is a placeholder for data preprocessing
    # In a real application, this would include:
    # - Handling missing values
    # - Feature engineering
    # - Normalization/scaling
    # - Temporal feature extraction
    
    processed_data = data.copy()
    
    # Placeholder operations
    if 'age' in processed_data.columns:
        processed_data['age_group'] = pd.cut(processed_data['age'], 
                                             bins=[0, 18, 40, 65, 100], 
                                             labels=['0-18', '19-40', '41-65', '65+'])
    
    # Add more preprocessing steps as needed
    
    return processed_data

def predict_risk(data):
    """
    Predict patient deterioration risk
    
    Parameters:
    data (pd.DataFrame): Preprocessed patient data
    
    Returns:
    dict: Prediction results including risk scores and explanations
    """
    # This is a placeholder for the risk prediction model
    # In a real application, this would:
    # - Load a trained model
    # - Generate predictions
    # - Calculate confidence intervals
    
    # Simulate predictions with random values
    num_patients = len(data)
    
    results = {
        'risk_scores': np.random.rand(num_patients) * 100,
        'confidence': np.random.rand(num_patients) * 0.3 + 0.7,  # 0.7-1.0 range
        'feature_importance': {
            'global': {
                'features': ['blood_glucose', 'medication_adherence', 'physical_activity', 
                            'blood_pressure', 'sleep_quality'],
                'importance': np.random.rand(5)
            },
            'local': {}
        },
        'metrics': {
            'auroc': 0.87,
            'auprc': 0.82,
            'calibration_score': 0.91
        }
    }
    
    # Generate local feature importance for each patient
    for i in range(num_patients):
        results['feature_importance']['local'][i] = {
            'features': ['blood_glucose', 'medication_adherence', 'physical_activity', 
                        'blood_pressure', 'sleep_quality'],
            'importance': np.random.rand(5)
        }
    
    return results

def generate_clinical_summary(patient_id, risk_score, feature_importance):
    """
    Generate a clinical summary of the risk prediction in natural language
    
    Parameters:
    patient_id (int): Patient identifier
    risk_score (float): Predicted risk score
    feature_importance (dict): Feature importance for the patient
    
    Returns:
    str: Natural language summary of risk prediction
    """
    # This is a placeholder for LLM-generated clinical summary
    # In a real application, this would:
    # - Use an LLM to generate clinician-friendly explanations
    # - Customize the summary based on patient-specific factors
    
    # Placeholder summary
    summary = f"""
    Patient {patient_id} shows {'elevated' if risk_score > 50 else 'moderate'} risk ({risk_score:.0f}%) 
    for deterioration within 90 days. 
    
    The primary contributing factors are:
    1. {'High blood glucose levels' if feature_importance[0] > 0.5 else 'Fluctuating blood glucose levels'}
    2. {'Poor medication adherence' if feature_importance[1] > 0.5 else 'Occasional medication non-adherence'}
    3. {'Declining physical activity' if feature_importance[2] > 0.5 else 'Inconsistent physical activity'}
    
    Recommended interventions include medication regimen review, 
    diabetes management education, and increased frequency of glucose monitoring.
    """
    
    return summary
