requirements.txt
streamlit==1.31.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
pillow==10.0.1
scikit-learn==1.3.0
